+++ src/XboxBatteryMonitor/Interfaces/INotificationService.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using Avalonia.Controls;
 
-namespace XboxBatteryMonitor.Services;
+namespace XboxBatteryMonitor.Interfaces;
 
 public interface INotificationService
 {
